Rename apiRoutes file to api
